fixup! src: consolidate exit codes in the code base

+++ src/env.cc
@@ -1521,8 +1521,9 @@ void Environment::Exit(ExitCode exit_code) {
               uv_os_getpid(), thread_id());
     }
 
-    fprintf(
-        stderr, "WARNING: Exited the environment with code %d\n", exit_code);
+    fprintf(stderr,
+            "WARNING: Exited the environment with code %d\n",
+            static_cast<int>(exit_code));
     PrintStackTrace(isolate(),
                     StackTrace::CurrentStackTrace(
                         isolate(), stack_trace_limit(), StackTrace::kDetailed));
